requirements.txt
requests
pandas
numpy
logging
pyupbit
python-dotenv
